src/requirements.txt
seaborn
geopandas
geodatasets
plotly